# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PveSphere is a comprehensive web-based management platform for Proxmox VE (PVE) clusters.",
        "title": "PveSphere API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "PveSphere Support",
            "url": "https://github.com/pvesphere/pvesphere",
            "email": "support@pvesphere.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/api/v1/vms": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE虚拟机模块"
                ],
                "summary": "获取虚拟机列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "集群名称",
                        "name": "cluster_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "节点名称",
                        "name": "node_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "应用ID",
                        "name": "app_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.ListVMResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "仅创建数据库记录，用于手动同步或导入场景，不调用 Proxmox API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE虚拟机模块"
                ],
                "summary": "创建虚拟机（仅数据库记录）",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.CreateVMRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/backup": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "使用 Proxmox vzdump API 创建虚拟机备份",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE虚拟机模块"
                ],
                "summary": "创建虚拟机备份",
                "parameters": [
                    {
                        "description": "创建备份请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.CreateBackupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.CreateBackupResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除指定存储中的备份文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE虚拟机模块"
                ],
                "summary": "删除虚拟机备份",
                "parameters": [
                    {
                        "description": "删除备份请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.DeleteBackupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.DeleteBackupResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/cloudinit": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取虚拟机的 CloudInit 配置，包含当前和待处理的值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE虚拟机模块"
                ],
                "summary": "获取虚拟机 CloudInit 配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "虚拟机ID",
                        "name": "vm_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.GetVMCloudInitResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "重新生成和更改 cloudinit 配置驱动器",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE虚拟机模块"
                ],
                "summary": "更新虚拟机 CloudInit 配置",
                "parameters": [
                    {
                        "description": "更新 CloudInit 配置请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.UpdateVMCloudInitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.UpdateVMCloudInitResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/config": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE虚拟机模块"
                ],
                "summary": "获取虚拟机当前配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "虚拟机ID",
                        "name": "vm_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.GetVMCurrentConfigResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE虚拟机模块"
                ],
                "summary": "更新虚拟机配置",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.UpdateVMConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.UpdateVMConfigResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/config/pending": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE虚拟机模块"
                ],
                "summary": "获取虚拟机pending配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "虚拟机ID",
                        "name": "vm_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.GetVMPendingConfigResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/console": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "调用 Proxmox vncproxy 获取 port/ticket 等信息，用于 noVNC 连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE虚拟机模块"
                ],
                "summary": "获取虚拟机 Console（VNCProxy）",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.GetVMConsoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.GetVMConsoleResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/console/ws": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "同域 WS 代理到 Proxmox vncwebsocket，供 noVNC 直接连接",
                "tags": [
                    "PVE虚拟机模块"
                ],
                "summary": "虚拟机 Console WebSocket（VNC WebSocket 代理）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ws_token（由 /api/v1/vms/console 返回）",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/vms/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "调用 Proxmox API 创建虚拟机并自动创建数据库记录，这是最常用的场景",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PVE虚拟机模块"
                ],
                "summary": "创建虚拟机（完整流程）",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.CreateVMRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pvesphere_api_v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pvesphere_api_v1.DeleteBackupResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "pvesphere_api_v1.ListVMResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/pvesphere_api_v1.ListVMResponseData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "pvesphere_api_v1.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}